Add a test, with the failing part commented out
I want to test that when:

- a one-hour event is loaded
- all-day is turned on

then the start and end date widgets display both the same date.

This fails right now, because we subtract one from the end_date.  I'll
change the way the dates are handled so that this magic only needs to
be done when setting back the GcalEvent.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-calendar/-/merge_requests/550>