gcal_schedule_section_day_changed() - Use the original dates
The original code from commit 10234653 is

  GCAL_RETURN (gcal_date_time_compare_date (start_date, gcal_event_get_date_start (self->event)) < 0 ||
               gcal_date_time_compare_date (end_date, gcal_event_get_date_end (self->event)) > 0);

but it doesn't explain why it just checks for the modified event
getting "wider" (earlier start, later end) than the original event.

I don't understand that logic, i.e. why doesn't it just check that the
dates changed.

Still, now that we keep the original date-time in the
GcalScheduleValues, use them to preserve the original semantics.

I'm not sure exactly what to test about this... a test to preserve those
semantics?  A test that the function, per its name, returns true if
the days are strictly different?

Part-of: <https://gitlab.gnome.org/GNOME/gnome-calendar/-/merge_requests/550>