Increase the required version of glib to 2.80.0
We'll need this for G_MUTEX_AUTO_LOCK().  It basically does

  g_autoptr (GMutexLocker) G_GNUC_UNUSED var = g_mutex_locker_new (mutex)

which will let us remove static-scan's warnings about assigned
variables that are never read, e.g. it complains about "locker" here:

    g_autoptr (GMutexLocker) locker = g_mutex_locker_new (&priv->shared.mutex);

The whole point of GMutexLocker is to be used with g_autoptr for a
mutex that is unlocked when its scope ends.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-calendar/-/merge_requests/581>